refactor(backup): move the nudge hook into the self-custodial namespace
`useBackupNudgeState` is self-custodial logic: every surface it returns is
gated on the account type, its dismissal keys are scoped to a self-custodial
account id, and it reads the self-custodial backup provider. It has no meaning
for a custodial account. Sitting in `app/hooks/` put account-type-specific
logic in the shared namespace, which is where the shared layer is supposed to
talk to both sides through an interface - `useActiveWallet` - rather than host
one side's concerns.

Moves, with the spec to the mirrored path, and exported through the namespace
index like its siblings:

  app/hooks/use-backup-nudge-state.ts
    -> app/self-custodial/hooks/use-backup-nudge-state.ts
  __tests__/hooks/use-backup-nudge-state.spec.ts
    -> __tests__/self-custodial/hooks/use-backup-nudge-state.spec.ts

The shared hooks it consumes - useAccountRegistry, useActiveWallet,
useIsSelfCustodialAccount - are genuinely account-type-agnostic and stay in
`app/hooks/`, now imported by path rather than as siblings.

No behaviour change; coverage stays at 100% statements, branches, functions
and lines. Neither unmerged recovery branch touches this file, so the move
does not conflict with them.
