+++ __tests__/screens/settings-screen/settings-screen.spec.tsx
@@ -1,4 +1,4 @@
-jest.mock("@app/hooks/use-backup-nudge-state", () => ({
+jest.mock("@app/self-custodial/hooks/use-backup-nudge-state", () => ({
   useBackupNudgeState: () => ({
     shouldShowBanner: false,
     shouldShowModal: false,
