+++ app/self-custodial/hooks/index.ts
@@ -1,3 +1,4 @@
+export { useBackupNudgeState } from "./use-backup-nudge-state"
 export { useNonCustodialConversionLimits } from "./use-non-custodial-conversion-limits"
 export { usePaymentRequest } from "./use-payment-request"
 export { usePendingDeposits } from "./use-pending-deposits"
